the port works when no uncaught exception is rethrown

+++ include/app/entities/BishopsBoard.hpp
@@ -26,7 +26,7 @@ class BishopsBoard : public AbstractBoardEntity
 		 * @return             [description]
 		 * 
 		 */
-		inline int identifyMoveType(bool xSuitColor, long oldPosition, long newPosition)
+		inline int identifyMoveType(bool xSuitColor, long long oldPosition, long long newPosition)
 		{
 
 		};
@@ -37,9 +37,9 @@ class BishopsBoard : public AbstractBoardEntity
 		 * @return              long
 		 * 
 		 */
-		inline long identifyMoveField(long old_position, long new_position)
+		inline long long identifyMoveField(long long old_position, long long new_position)
 		{
-			long historyOfMove = old_position | new_position;
+			long long historyOfMove = old_position | new_position;
 
 			std::cout << "historyOfMove" << std::endl;
 			HelperFunctions::drawArrayBoardFromBitBoard(historyOfMove);
@@ -47,8 +47,8 @@ class BishopsBoard : public AbstractBoardEntity
 			/* is it left diagonal? */
 			for (int i = 0; i < 15; ++i)
 			{
-				long temp_one = BishopsBoard::left_diagonals[i] | historyOfMove;
-				long temp_two = temp_one ^BishopsBoard::left_diagonals[i];
+				long long temp_one = BishopsBoard::left_diagonals[i] | historyOfMove;
+				long long temp_two = temp_one ^BishopsBoard::left_diagonals[i];
 			    if( temp_two == 0)
 			   		return BishopsBoard::left_diagonals[i];
 			}
@@ -56,8 +56,8 @@ class BishopsBoard : public AbstractBoardEntity
 			/* is it right diagonal? */
 			for (int i = 0; i < 15; ++i)
 			{
-				long temp_one = BishopsBoard::right_diagonals[i] | historyOfMove;
-				long temp_two = temp_one^ BishopsBoard::right_diagonals[i];
+				long long temp_one = BishopsBoard::right_diagonals[i] | historyOfMove;
+				long long temp_two = temp_one^ BishopsBoard::right_diagonals[i];
 				if( temp_two  == 0)
 					return BishopsBoard::right_diagonals[i];
 			}
@@ -72,12 +72,12 @@ class BishopsBoard : public AbstractBoardEntity
 		 * @return              [description]
 		 * 
 		 */
-		inline long identifyMoveDirection(long old_position, long new_position)
+		inline long long identifyMoveDirection(long long old_position, long long new_position)
 		{
 			return new_position - old_position;
 		};
 
-		inline std::string getMoveDirection(long old_position, long new_position)
+		inline std::string getMoveDirection(long long old_position, long long new_position)
 		{
 			std::string upleft    = "up_left";
 			std::string downleft  = "down_left";
@@ -85,8 +85,8 @@ class BishopsBoard : public AbstractBoardEntity
 			std::string downright = "down_right";
 			std::string invalid   = "invalid";
 
-			long direction    = this->identifyMoveDirection(old_position, new_position);
-			long rank_or_file = this->identifyMoveField(old_position, new_position);
+			long long direction    = this->identifyMoveDirection(old_position, new_position);
+			long long rank_or_file = this->identifyMoveField(old_position, new_position);
 
 			std::cout << "rank_or_file: " << rank_or_file << std::endl;
 
@@ -132,12 +132,12 @@ class BishopsBoard : public AbstractBoardEntity
 		 * @return               [description]
 		 * 
 		 */
-		inline int identifyMoveType(long old_position, long new_position, long fullboard, long moveDirection)
+		inline int identifyMoveType(long long old_position, long long new_position, long long fullboard, long long moveDirection)
 		{
 
 		};
 
-		inline int getLeftDiagonalOf(long position)
+		inline int getLeftDiagonalOf(long long position)
 		{
 			for (int i = 0; i < 15; ++i)
 				if( ((BishopsBoard::left_diagonals[i] & position) ^ position) == 0 ) return i;
@@ -145,7 +145,7 @@ class BishopsBoard : public AbstractBoardEntity
 			return -1;
 		};
 
-		inline int getRightDiagonalOf(long position)
+		inline int getRightDiagonalOf(long long position)
 		{
 			for (int i = 0; i < 15; ++i)
 				if( ((BishopsBoard::right_diagonals[i] & position) ^ position) == 0 ) return i;
@@ -161,12 +161,12 @@ class BishopsBoard : public AbstractBoardEntity
 		 * @return              [description]
 		 * 
 		 */
-		inline long getBoardBetweenPositions(long old_position, long new_position, long fullboard, long field, std::string direction)
+		inline long long getBoardBetweenPositions(long long old_position, long long new_position, long long fullboard, long long field, std::string direction)
 		{
-			long board_old_and_new         = old_position | new_position;
-			long fullboard_on_moving_field = HelperFunctions::applyMask_Keep(fullboard, field);
+			long long board_old_and_new         = old_position | new_position;
+			long long fullboard_on_moving_field = HelperFunctions::applyMask_Keep(fullboard, field);
 
-			long mask = 0L;
+			long long mask = 0L;
 
 			if( direction.compare("up_left") == 0 || direction.compare("down_left") == 0)
 			{
@@ -184,7 +184,7 @@ class BishopsBoard : public AbstractBoardEntity
 			}
 			else if( direction.compare("up_right") == 0 || direction.compare("down_right") == 0)
 			{
-				long file_one = this->getLeftDiagonalOf(old_position), 
+				long long file_one = this->getLeftDiagonalOf(old_position), 
 					 file_two = this->getLeftDiagonalOf(new_position);
 
 
@@ -226,11 +226,11 @@ class BishopsBoard : public AbstractBoardEntity
 		 * @return [description]
 		 * 
 		 */
-		inline bool isBoardEmptyBetweenOldAndNew(long old_position, long new_position, long fullboard, long move_direction, std::string direction)
+		inline bool isBoardEmptyBetweenOldAndNew(long long old_position, long long new_position, long long fullboard, long long move_direction, std::string direction)
 		{
 			if( direction.compare("invalid") == 0 ) return false;
 
-			long what_in_between = this->getBoardBetweenPositions(old_position, new_position, fullboard, move_direction, direction);
+			long long what_in_between = this->getBoardBetweenPositions(old_position, new_position, fullboard, move_direction, direction);
 
 			if( what_in_between == 0) return true;
 			else return false;
@@ -238,14 +238,14 @@ class BishopsBoard : public AbstractBoardEntity
 
 		/**
 		 * Judges the new move as valid or invalid.
-		 * @param  long : the new board's configuration that represents the new move.
+		 * @param  long long : the new board's configuration that represents the new move.
 		 * @return bool : true upon valid.
 		 *  
 		 */
-		inline bool isMoveValid(long move, long fullboard, long myFriendsBoard)
+		inline bool isMoveValid(long long move, long long fullboard, long long myFriendsBoard)
 		{
-			long oldPosition = this->extractOldPosition(move);
-			long newPosition = this->extractNewMove(move);
+			long long oldPosition = this->extractOldPosition(move);
+			long long newPosition = this->extractNewMove(move);
 
 			if(0 == newPosition) return false; // If there was no new move
 
@@ -300,7 +300,7 @@ class BishopsBoard : public AbstractBoardEntity
 		 * Alters the positions at this board.
 		 * 
 		 */
-		inline void alterBoard(bool isInitLoad, long fullboard, long myFriendsBoard, std::string& binaryString)
+		inline void alterBoard(bool isInitLoad, long long fullboard, long long myFriendsBoard, std::string& binaryString)
 		{
 			AbstractBoardEntity::alterBoard(isInitLoad, fullboard, myFriendsBoard, binaryString);
 		};
