+++ include/app/entities/KnightsBoard.hpp
@@ -27,16 +27,16 @@ class KnightsBoard : public AbstractBoardEntity
 		 * @return             [description]
 		 * 
 		 */
-		inline int identifyMoveType(bool xSuitColor, long oldPosition, long newPosition)
+		inline int identifyMoveType(bool xSuitColor, long long oldPosition, long long newPosition)
 		{
 
 		};
 
-		inline int identifyMoveType(bool xSuitColor, long oldPosition, long newPosition, long fullboard)
+		inline int identifyMoveType(bool xSuitColor, long long oldPosition, long long newPosition, long long fullboard)
 		{
 			if(0 == newPosition) return 0; // If there was no new move
 
-			long special_case = KnightsBoard::right_diagonals[14];
+			long long special_case = KnightsBoard::right_diagonals[14];
 
 			bool moveForwardRightL, moveForwardLeftL, moveForwardLateralRightL, moveForwardLateralLeftL,
 				 moveBackwardRightL, moveBackwardLeftL, moveBackwardLateralRightL, moveBackwardLateralLeftL;
@@ -55,7 +55,7 @@ class KnightsBoard : public AbstractBoardEntity
 			}
 			else
 			{
-				long lsb_container;
+				long long lsb_container;
 
 				lsb_container      = oldPosition>>17;
 				lsb_container      = lsb_container<<(64-17);
@@ -88,14 +88,14 @@ class KnightsBoard : public AbstractBoardEntity
 
 		/**
 		 * Judges the new move as valid or invalid.
-		 * @param  long : the new board's configuration that represents the new move.
+		 * @param  long long : the new board's configuration that represents the new move.
 		 * @return bool : true upon valid.
 		 *  
 		 */
-		inline bool isMoveValid(long move, long fullboard, long myFriendsBoard)
+		inline bool isMoveValid(long long move, long long fullboard, long long myFriendsBoard)
 		{
-			long oldPosition = this->extractOldPosition(move);
-			long newPosition = this->extractNewMove(move);
+			long long oldPosition = this->extractOldPosition(move);
+			long long newPosition = this->extractNewMove(move);
 
 			int moveType = this->identifyMoveType(this->suitColor, oldPosition, newPosition, fullboard);
 			if(moveType == 1) return true;
@@ -130,7 +130,7 @@ class KnightsBoard : public AbstractBoardEntity
 		 * Alters the positions at this board.
 		 * 
 		 */
-		inline void alterBoard(bool isInitLoad, long fullboard, long myFriendsBoard, std::string& binaryString)
+		inline void alterBoard(bool isInitLoad, long long fullboard, long long myFriendsBoard, std::string& binaryString)
 		{
 			AbstractBoardEntity::alterBoard(isInitLoad, fullboard, myFriendsBoard, binaryString);
 		};
